refactor(codegen): move emitUnionTypeAnnotation into parsers-primitives
