+++ packages/react-native-codegen/src/parsers/parsers-primitives.js
@@ -25,6 +25,7 @@ import type {
   NativeModuleNumberTypeAnnotation,
   NativeModulePromiseTypeAnnotation,
   NativeModuleTypeAliasTypeAnnotation,
+  NativeModuleUnionTypeAnnotation,
   ObjectTypeAnnotation,
   ReservedTypeAnnotation,
   StringTypeAnnotation,
@@ -38,6 +39,7 @@ import type {
   TypeDeclarationMap,
 } from './utils';
 
+const {UnsupportedUnionTypeAnnotationParserError} = require('./errors');
 const {
   throwIfArrayElementTypeAnnotationIsUnsupported,
 } = require('./error-utils');
@@ -249,6 +251,32 @@ function emitFloat(
   });
 }
 
+function emitUnionTypeAnnotation(
+  nullable: boolean,
+  hasteModuleName: string,
+  typeAnnotation: $FlowFixMe,
+  parser: Parser,
+): Nullable<NativeModuleUnionTypeAnnotation> {
+  const unionTypes = parser.remapUnionTypeAnnotationMemberNames(
+    typeAnnotation.types,
+  );
+
+  // Only support unionTypes of the same kind
+  if (unionTypes.length > 1) {
+    throw new UnsupportedUnionTypeAnnotationParserError(
+      hasteModuleName,
+      typeAnnotation,
+      unionTypes,
+      parser.language(),
+    );
+  }
+
+  return wrapNullable(nullable, {
+    type: 'UnionTypeAnnotation',
+    memberType: unionTypes[0],
+  });
+}
+
 function translateArrayTypeAnnotation(
   hasteModuleName: string,
   types: TypeDeclarationMap,
@@ -319,6 +347,7 @@ module.exports = {
   emitString,
   emitStringish,
   emitMixedTypeAnnotation,
+  emitUnionTypeAnnotation,
   typeAliasResolution,
   translateArrayTypeAnnotation,
 };
