+++ packages/react-native-codegen/src/parsers/typescript/modules/index.js
@@ -28,7 +28,6 @@ const {resolveTypeAnnotation, getTypes} = require('../utils.js');
 const {
   assertGenericTypeAnnotationHasExactlyOneTypeParameter,
   parseObjectProperty,
-  emitUnionTypeAnnotation,
   translateDefault,
   buildPropertySchema,
 } = require('../../parsers-commons');
@@ -47,6 +46,7 @@ const {
   emitString,
   emitStringish,
   emitMixedTypeAnnotation,
+  emitUnionTypeAnnotation,
   typeAliasResolution,
   translateArrayTypeAnnotation,
 } = require('../../parsers-primitives');
